pkg/aggregator/tags/tlm_test.go
// Unless explicitly stated otherwise all files in this repository are licensed
// under the Apache License Version 2.0.
// This product includes software developed at Datadog (https://www.datadoghq.com/).
// Copyright 2021-present Datadog, Inc.

package tags

import (
	"testing"

	"github.com/DataDog/datadog-agent/pkg/tagset"
	"github.com/stretchr/testify/require"
)

func TestStore(t *testing.T) {
	c := NewTlm(true, "test")

	t1 := tagset.NewTags([]string{"1"})
	h1 := t1.Hash()
	t2 := tagset.NewTags([]string{"2"})
	h2 := t2.Hash()

	c.Use(t1)

	require.EqualValues(t, 1, len(c.tagsByKey))
	require.EqualValues(t, 1, c.cap)
	require.EqualValues(t, 1, c.tagsByKey[h1].refs)

	c.Use(t1)
	require.EqualValues(t, 1, len(c.tagsByKey))
	require.EqualValues(t, 1, c.cap)
	require.EqualValues(t, 2, c.tagsByKey[h1].refs)

	c.Use(t2)
	require.EqualValues(t, 2, len(c.tagsByKey))
	require.EqualValues(t, 2, c.cap)
	require.EqualValues(t, 2, c.tagsByKey[h1].refs)
	require.EqualValues(t, 1, c.tagsByKey[h2].refs)

	c.Use(t2)
	require.EqualValues(t, 2, len(c.tagsByKey))
	require.EqualValues(t, 2, c.cap)
	require.EqualValues(t, 2, c.tagsByKey[h1].refs)
	require.EqualValues(t, 2, c.tagsByKey[h2].refs)

	c.Release(t1)
	require.EqualValues(t, 2, len(c.tagsByKey))
	require.EqualValues(t, 2, c.cap)
	require.EqualValues(t, 1, c.tagsByKey[h1].refs)
	require.EqualValues(t, 2, c.tagsByKey[h2].refs)

	c.Shrink()
	require.EqualValues(t, 2, len(c.tagsByKey))
	require.EqualValues(t, 2, c.cap)

	c.Release(t2)
	require.EqualValues(t, 2, len(c.tagsByKey))
	require.EqualValues(t, 2, c.cap)
	require.EqualValues(t, 1, c.tagsByKey[h1].refs)
	require.EqualValues(t, 1, c.tagsByKey[h2].refs)

	c.Release(t1)
	require.EqualValues(t, 2, len(c.tagsByKey))
	require.EqualValues(t, 2, c.cap)
	require.EqualValues(t, 0, c.tagsByKey[h1].refs)
	require.EqualValues(t, 1, c.tagsByKey[h2].refs)

	c.Shrink()
	require.EqualValues(t, 1, len(c.tagsByKey))
	require.EqualValues(t, 2, c.cap)
	require.EqualValues(t, 1, c.tagsByKey[h2].refs)

	c.Release(t2)
	require.EqualValues(t, 1, len(c.tagsByKey))
	require.EqualValues(t, 2, c.cap)
	require.EqualValues(t, 0, c.tagsByKey[h2].refs)

	c.Shrink()
	require.EqualValues(t, 0, len(c.tagsByKey))
	require.EqualValues(t, 0, c.cap)
}

func TestStoreDisabled(t *testing.T) {
	c := NewTlm(false, "test")

	t1 := tagset.NewTags([]string{"1"})
	t2 := tagset.NewTags([]string{"2"})

	c.Use(t1)
	require.EqualValues(t, 0, len(c.tagsByKey))
	require.EqualValues(t, 0, c.cap)

	c.Use(t1)
	require.EqualValues(t, 0, len(c.tagsByKey))
	require.EqualValues(t, 0, c.cap)

	c.Use(t2)
	require.EqualValues(t, 0, len(c.tagsByKey))
	require.EqualValues(t, 0, c.cap)

	c.Release(t1)
	require.EqualValues(t, 0, len(c.tagsByKey))
	require.EqualValues(t, 0, c.cap)

	c.Release(t2)
	require.EqualValues(t, 0, len(c.tagsByKey))
	require.EqualValues(t, 0, c.cap)

	c.Shrink()
	require.EqualValues(t, 0, len(c.tagsByKey))
	require.EqualValues(t, 0, c.cap)
}
